Support multiple duplicate queryparams
The SendGrid API documentation states that the user can supply up to ten
categories, but there is no way to supply them using the Java library.

This commit fixes the problem by interpreting any value that has an
ampersand in it as a list of values, which will be inserted as multiple
query params.

For example, the call request.addQueryParam("categories", "cat1&cat2")
will output "?categories=cat1&categories=cat2" into the URL.